I am sorry Eloy to submit this commit without comments but they should be as long as a book.
I made the changes I already told you plus much more but I strongly feel this is the right way to proceed in the development, as widely discussed with colleagues and statistical.
The diff file is 9 thousand row long.
Let's say this is the release I am going to fix and clean up to the first public release.
There should be a lot of errors because my testing, although long, was nothing against the size of the monster.