added management of plugins following the example of the assign module
In the frame of this work I also:
commented old style work in settings.php
found the following issues:

===============================================================
adminmanageplugins.php: issue 1
===============================================================
if in
http://localhost/HEAD/mod/survey/adminmanageplugins.php?subtype=surveyfield
you push the 'settings' link in the row of age plugin, you get an error. Why?

===============================================================
adminmanageplugins.php: issue 2
===============================================================
'order' and 'delete' in
I like to be coherent with all the other code but what can I do with them?

===============================================================
adminmanageplugins.php: note
===============================================================
In the frame of this work, I also provided an answer to what Joseph Rezeau asked me privately in an email

From Joseph Rezeau email:
-> * Note: it would be useful to have a "manage question types" feature for
this module, similar to the Quiz questions manage question types, so
that an Admin would be able to disable/enable survey question types at
site level. (thanks to JR)